+++ elastalert/alerters/teams.py
@@ -19,6 +19,8 @@ class MsTeamsAlerter(Alerter):
         self.ms_teams_alert_summary = self.rule.get('ms_teams_alert_summary', 'ElastAlert Message')
         self.ms_teams_alert_fixed_width = self.rule.get('ms_teams_alert_fixed_width', False)
         self.ms_teams_theme_color = self.rule.get('ms_teams_theme_color', '')
+        self.ms_teams_ca_certs = self.rule.get('ms_teams_ca_certs')
+        self.ms_teams_ignore_ssl_errors = self.rule.get('ms_teams_ignore_ssl_errors', False)
 
     def format_body(self, body):
         if self.ms_teams_alert_fixed_width:
@@ -32,6 +34,14 @@ class MsTeamsAlerter(Alerter):
         body = self.format_body(body)
         # post to Teams
         headers = {'content-type': 'application/json'}
+    
+        if self.ms_teams_ca_certs:
+            verify = self.ms_teams_ca_certs
+        else:
+            verify = not self.ms_teams_ignore_ssl_errors
+        if self.ms_teams_ignore_ssl_errors:
+            requests.packages.urllib3.disable_warnings()
+
         # set https proxy, if it was provided
         proxies = {'https': self.ms_teams_proxy} if self.ms_teams_proxy else None
         payload = {
@@ -46,7 +56,8 @@ class MsTeamsAlerter(Alerter):
 
         for url in self.ms_teams_webhook_url:
             try:
-                response = requests.post(url, data=json.dumps(payload, cls=DateTimeEncoder), headers=headers, proxies=proxies)
+                response = requests.post(url, data=json.dumps(payload, cls=DateTimeEncoder),
+                                         headers=headers, proxies=proxies, verify=verify)
                 response.raise_for_status()
             except RequestException as e:
                 raise EAException("Error posting to ms teams: %s" % e)